requirements.txt
tqdm
requests
getch
packaging
Pillow==9.5.0
opencv-python
scikit-image
scikit-learn
numpy
matplotlib
scipy==1.9.1
protobuf~=3.20
tsp_solver
gphoto2==2.5.0
pyrealsense2
tensorboardX==2.1
torch 
torchvision
torchgeometry
kornia
colour-science
ftfy
regex
git+https://github.com/openai/CLIP.git
librosa
timm
transformers
<<<<<<< HEAD
diffusers==0.18.1
=======
diffusers
>>>>>>> ffd451483ca3dcd3e79e4fd838237c30c732a213
accelerate
transformers
whisper
xarm
